FAD-4372 Reorganize helpers
